Improve diner_loop with mutex, need to check dead

+++ get_msec_since_start.c
@@ -6,13 +6,13 @@
 /*   By: tsierra- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/09 16:19:27 by tsierra-          #+#    #+#             */
-/*   Updated: 2021/07/09 16:19:50 by tsierra-         ###   ########.fr       */
+/*   Updated: 2021/07/12 18:08:10 by tsierra-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-unsigned long	get_msec_since_start(unsigned long start_time)
+unsigned	get_msec_since_start(unsigned long start_time)
 {
 	unsigned long	current_time;
